requirements.txt
fastapi
uvicorn
pydantic
groq
pymongo
bcrypt
python-dotenv
duckduckgo-search
langdetect
deep-translator
